requirements.txt
streamlit
psycopg2-binary
python-dotenv
PyYAML
requests
pytest